requirements.txt
requests==2.31.0
flask==2.3.3
flask-socketio==5.3.6
flask-cors==4.0.0
python-dotenv==1.0.0
pyjwt==2.8.0
werkzeug==2.3.7
psycopg2-binary==2.9.7
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
factory-boy==3.3.0
click==8.1.7
alembic==1.12.1
